Image Annotations (1/2)
Image annotations for ECB+ topics 1-36.
Each folder corresponds to an ECB+ article.
ECB folders may contain found_url.txt which is a text file with a single line containing the URL I located for the text via Google search.
ECB+ folders may contain broken_url.txt which is either empty (indicating the URL was broken and no archived version was able to be located) or contains a link to the archived article.
For each image, if an image caption was present, there is a corresponding .txt file with the same name as the image file.
Image file names were preserved if present.